Refactor maction, menclose, semantics and Notation utility into common and chtml components
